requirements.txt


torch
torchvision
numpy
matplotlib
nibabel
albumentations
segmentation-models-pytorch
flask
pillow
pytorch-grad-cam
scikit-learn
shap
tqdm
